requirements.txt
Package                  Version
------------------------ -----------
aiohttp                  3.8.4
aiosignal                1.3.1
alembic                  1.10.4
altair                   5.0.1
anyio                    3.6.2
appnope                  0.1.3
argon2-cffi              21.3.0
argon2-cffi-bindings     21.2.0
arrow                    1.2.3
asttokens                2.2.1
async-timeout            4.0.2
attrs                    23.1.0
backcall                 0.2.0
beautifulsoup4           4.12.2
bleach                   6.0.0
blessed                  1.20.0
blinker                  1.6.2
bs4                      0.0.1
cachetools               5.3.1
certifi                  2023.5.7
cffi                     1.15.1
charset-normalizer       3.1.0
click                    8.1.3
cloudpickle              2.2.1
comm                     0.1.3
contourpy                1.0.7
croniter                 1.3.14
cycler                   0.11.0
databricks-cli           0.17.7
dateutils                0.6.12
debugpy                  1.6.7
decorator                5.1.1
deepdiff                 6.3.0
defusedxml               0.7.1
distlib                  0.3.6
docker                   6.1.2
entrypoints              0.4
executing                1.2.0
fastapi                  0.88.0
fastjsonschema           2.16.3
filelock                 3.12.0
Flask                    2.3.2
fonttools                4.39.4
fqdn                     1.5.1
frozenlist               1.3.3
fsspec                   2023.5.0
gitdb                    4.0.10
GitPython                3.1.31
grpcio                   1.49.1
gunicorn                 20.1.0
h11                      0.14.0
idna                     3.4
importlib-metadata       6.6.0
importlib-resources      5.12.0
inquirer                 3.1.3
ipykernel                6.23.0
ipython                  8.13.2
ipython-genutils         0.2.0
isoduration              20.11.0
itsdangerous             2.1.2
jedi                     0.18.2
Jinja2                   3.1.2
joblib                   1.2.0
jsonpointer              2.3
jsonschema               4.17.3
jupyter_client           8.2.0
jupyter_core             5.3.0
jupyter-events           0.6.3
jupyter_server           2.5.0
jupyter_server_terminals 0.4.4
jupyterlab-pygments      0.2.2
kiwisolver               1.4.4
lightning                2.0.2
lightning-cloud          0.5.34
lightning-utilities      0.8.0
Mako                     1.2.4
Markdown                 3.4.3
markdown-it-py           2.2.0
MarkupSafe               2.1.2
matplotlib               3.7.1
matplotlib-inline        0.1.6
mdurl                    0.1.2
mistune                  2.0.5
mlflow                   2.3.2
mpmath                   1.3.0
msgpack                  1.0.5
multidict                6.0.4
nbclassic                1.0.0
nbclient                 0.7.4
nbconvert                7.4.0
nbformat                 5.8.0
nest-asyncio             1.5.6
networkx                 3.1
notebook                 6.5.4
notebook_shim            0.2.3
numpy                    1.24.3
oauthlib                 3.2.2
opencv-python            4.7.0.72
ordered-set              4.1.0
packaging                23.1
pandas                   2.0.1
pandocfilters            1.5.0
parso                    0.8.3
pexpect                  4.8.0
pickleshare              0.7.5
Pillow                   9.5.0
pip                      23.1.2
platformdirs             3.5.0
prometheus-client        0.16.0
prompt-toolkit           3.0.38
protobuf                 4.23.0
psutil                   5.9.5
ptyprocess               0.7.0
pure-eval                0.2.2
pyarrow                  11.0.0
pycparser                2.21
pydantic                 1.10.7
pydeck                   0.8.1b0
Pygments                 2.15.1
PyJWT                    2.6.0
Pympler                  1.0.1
pyparsing                3.0.9
pyrsistent               0.19.3
python-dateutil          2.8.2
python-editor            1.0.4
python-json-logger       2.0.7
python-multipart         0.0.6
pytorch-lightning        2.0.2
pytz                     2023.3
pytz-deprecation-shim    0.1.0.post0
PyYAML                   6.0
pyzmq                    25.0.2
querystring-parser       1.2.4
ray                      2.4.0
readchar                 4.0.5
requests                 2.30.0
rfc3339-validator        0.1.4
rfc3986-validator        0.1.1
rich                     13.3.5
scikit-learn             1.2.2
scipy                    1.10.1
seaborn                  0.12.2
Send2Trash               1.8.2
setuptools               67.7.2
six                      1.16.0
smmap                    5.0.0
sniffio                  1.3.0
soupsieve                2.4.1
SQLAlchemy               2.0.13
sqlparse                 0.4.4
stack-data               0.6.2
starlette                0.22.0
starsessions             1.3.0
streamlit                1.23.1
sympy                    1.11.1
tabulate                 0.9.0
tenacity                 8.2.2
terminado                0.17.1
threadpoolctl            3.1.0
tinycss2                 1.2.1
toml                     0.10.2
toolz                    0.12.0
torch                    2.0.1
torchmetrics             0.11.4
torchvision              0.15.2
tornado                  6.3.1
tqdm                     4.65.0
traitlets                5.9.0
typing_extensions        4.5.0
tzdata                   2023.3
tzlocal                  4.3
uri-template             1.2.0
urllib3                  1.26.15
uvicorn                  0.22.0
validators               0.20.0
virtualenv               20.21.0
wcwidth                  0.2.6
webcolors                1.13
webencodings             0.5.1
websocket-client         1.5.1
websockets               11.0.3
Werkzeug                 2.3.4
wheel                    0.40.0
yarl                     1.9.2
zipp                     3.15.0
